requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
plotly>=5.15.0
altair>=5.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
requests>=2.31.0
urllib3>=2.0.0
